+++ src/kitcube-devices/simrandom.h
@@ -36,7 +36,7 @@ public:
 	
 	void writeHeader();
 	
-	void parseData(char *line, struct timeval *l_tData, float *sensorValue);
+	void parseData(char *line, struct timeval *l_tData, double *sensorValue);
 	
 	void writeData();
